Refactor messaging system for more advanced messaging.
Now using a unique identifier rather than the mac address. This will bring pros and cons. Basically Im going to have a one time use chat that needs to be set
every time you use the app. This will provide saftey but no real data persistance. Im still contimplating how messaging will be refactored.

I could send the message directly to the senders system, or I could use a data base. The issue with a database is that I will be storing data, and to be honest I want to minimize
the overhead of data storage. A, this should be free and easy to use. Im not google I cant freely store data whereever. Because of this I may transition to using something like a temporary cache on each system that we query and some form of communication through other methods.

Who knows. Ill do more research